Announce New Users
Updated client.js to announce new users connecting or disconnecting and listing the count of users. Updated server.js to announce new users connecting or disconnecting and listing count of users. Also updated PORT in server.js as it should have been done back in "Set up the Environment." This change doesn't impact what info is shown, just how it's accessed, which is why it didn't fail the previous tests.